Beginnings of having this library render HTML.
